added support for optional running of one ets table per module.
by setting config :memoize, caches: [ ModuleName ] in your config file
you can make use of this feature. it is backwards compatible with the
old way of doing things, and forked the default code leaving behind a
simple module that will ignore these changes as best it can.  I still
need to fix evection  but will get to that soon.